+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## Changelog ##
 
 ### v1.8 (in development) - (Dev time: 1 hour) ###
+- Tweak: Format email content
 - Tweak: Remove PHP4 compatible code
 
 ### v1.7.5  (2014-09-23) - (Dev time: 1 hour) ###